custom-packages-requirements.txt
## MeCab
mecab-python3

## gensim
Cython
boto
google-compute-engine
gensim

## spaCy
spacy

## GCloud Natural Language API
google-cloud-language

## GCloud Storage
google-cloud-storage
